+++ lib/wnm_support/active_record_ext/multi_action.rb
@@ -1,0 +1,13 @@
+module WnmSupport
+  module ActiveRecordExt
+    module MultiAction
+      extend ActiveSupport::Concern
+
+      module ClassMethods
+        def multi
+          [:destroy]
+        end
+      end
+    end
+  end
+end