+++ lib/wnm_support/active_record_ext/view_helpers.rb
@@ -1,0 +1,9 @@
+module WnmSupport
+  module ActiveRecordExt
+    module ViewHelpers
+      def helpers
+        #ActionController::Base.helpers
+      end
+    end
+  end
+end