+++ spec/wnm_support/activer_record_ext/view_helpers_spec.rb
@@ -1,0 +1,11 @@
+require "spec_helper"
+
+class Item < ActiveRecord::Base
+  include ActiveRecordNoTable
+end
+
+describe WnmSupport::ActiveRecordExt::ViewHelpers do
+  it "should add helpers method to each model" do
+    Item.new.should respond_to :helpers
+  end
+end