+++ changelog.md
@@ -1,3 +1,14 @@
+# 0.0.3
+* add support for active record extentions
+** DestroyValidation
+** DestroyValidation
+** ViewHelpers
+** MultiAction
+** MysqlTruncate
+** MysqlOrderByField
+* add mysql connection with one testing model
+** news name:string
+
 # 0.0.2
 * rename gem to wnm_support